# Documentation/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: ECMA Events
  description: >-
    API da aplicação gerida pela companhia de marketing ECMA para a gestão da
    organização de eventos com a colaboração das suas companhias parceiras.
servers:
  - url: 'http://localhost:8080/'
...
paths:
...
  /proposals:
    get:
      tags:
        - Event Proposal
      description: Obter todas as propostas de eventos do sistema
      operationId: getEventProposals
      parameters:
        - name: criteria
          in: query
          description: Filtrar as propostas de eventos que satisfazem o critério
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Lista com a informação sobre as propostas de eventos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventProposal'
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Event Proposal
      description: Adicionar uma proposta de evento ao sistema
      operationId: addEventProposal
      requestBody:
        description: Todos os necessários dados da proposta de evento
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventProposal'
      responses:
        '200':
          description: Proposta de evento adicionada com sucesso
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/proposals/{id}':
    get:
      tags:
        - Event Proposal
      description: 'Obter a proposta de evento identificado pelo {id}'
      operationId: getEventProposal
      parameters:
        - name: id
          in: path
          description: Identificador da proposta de evento a obter
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Informação da proposta de evento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventProposal'
        '404':
          description: Proposta de evento não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Event Proposal
      description: 'Atualizar a proposta de evento identificada pelo {id}'
      operationId: updateEventProposal
      parameters:
        - name: id
          in: path
          description: Identificador da proposta de evento a remover
          required: true
          schema:
            type: integer
            format: int64
        - name: proposal
          in: query
          description: Objeto proposta de evento que é atualizado
          required: true
          schema:
            $ref: '#/components/schemas/EventProposal'
      responses:
        '200':
          description: Proposta de evento atualizada
        '400':
          description: Id especificado inválido
        '404':
          description: Proposta de evento não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Event Proposal
      description: 'Remover do sistema a proposta de evento identificada pelo {id}'
      operationId: deleteEventProposal
      parameters:
        - name: id
          in: path
          description: Identificador da proposta de evento a remover
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Proposta de evento removida
        '404':
          description: Proposta de evento não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/proposals/{id}/sections':
    get:
      tags:
        - Event Proposal
      description: 'Obter todos as secçõe da proposta de evento identificada pelo {id}'
      operationId: getEventProposalSections
      parameters:
        - name: id
          in: path
          description: Identificador da proposta de evento
          required: true
          schema:
            type: integer
            format: int64
        - name: criteria
          in: query
          description: Filtrar as secções que satisfazem o critério
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Lista com as secções da proposta de evento em questão
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventProposalSection'
        '404':
          description: Proposta de evento não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Event Proposal
      description: 'Adicionar uma secção à proposta de evento identificada pelo {id}'
      operationId: addEventProposalSection
      parameters:
        - name: id
          in: path
          description: Identificador da proposta de evento
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Todos os necessários dados da secção
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventProposalSection'
      responses:
        '200':
          description: Secção adicionada com sucesso
        '404':
          description: Proposta de evento não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/proposals/{pid}/sections/{sid}':
    put:
      tags:
        - Event Proposal
      description: 'Atualizar a secção identificada pelo {sid} na proposta {pid}'
      operationId: updateProposalSection
      parameters:
        - name: pid
          in: path
          description: Identificador da proposta
          required: true
          schema:
            type: integer
            format: int64
        - name: sid
          in: path
          description: Identificador da secção a atualizar
          required: true
          schema:
            type: integer
            format: int64
        - name: section
          in: query
          description: Objeto section que é atualizado
          required: true
          schema:
            $ref: '#/components/schemas/EventProposalSection'
      responses:
        '200':
          description: Secção atualizada
        '400':
          description: Id especificado da proposta ou secção inválido
        '404':
          description: Proposta ou secção não encontrados
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Event Proposal
      description: 'Remover do sistema a secção identificada pelo {sid} da proposta {pid}'
      operationId: deleteProposalSection
      parameters:
        - name: pid
          in: path
          description: Identificador da proposta
          required: true
          schema:
            type: integer
            format: int64
        - name: sid
          in: path
          description: Identificador da secção a remover
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Secção removida
        '404':
          description: Proposta ou secção não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/proposals/{id}/reviews':
    get:
      tags:
        - Event Proposal
      description: 'Obter todas as revisões da proposta de evento identificada pelo {id}'
      operationId: getEventProposalReviews
      parameters:
        - name: id
          in: path
          description: Identificador da proposta de evento
          required: true
          schema:
            type: integer
            format: int64
        - name: criteria
          in: query
          description: Filtrar as revisões que satisfazem o critério
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Lista com as revisões da proposta de evento em questão
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventProposalReview'
        '404':
          description: Proposta de evento não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Event Proposal
      description: 'Adicionar uma revisão à proposta de evento identificada pelo {id}'
      operationId: addEventProposalReview
      parameters:
        - name: id
          in: path
          description: Identificador da proposta de evento
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Todos os necessários dados da revisão
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventProposalReview'
      responses:
        '200':
          description: Revisão adicionada com sucesso
        '404':
          description: Proposta de evento não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/proposals/{pid}/reviews/{rid}':
    get:
      tags:
        - Event Proposal
      description: 'Obter a revisão identificada pelo {rid} da proposta {pid}'
      operationId: getProposalReview
      parameters:
        - name: pid
          in: path
          description: Identificador da proposta
          required: true
          schema:
            type: integer
            format: int64
        - name: rid
          in: path
          description: Identificador da revisão a obter
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Informação da revisão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventProposalReview'
        '404':
          description: Proposta ou revisão não encontradas
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Event Proposal
      description: 'Atualizar a revisão identificada pelo {rid} na empresa {pid}'
      operationId: updateProposalReview
      parameters:
        - name: pid
          in: path
          description: Identificador da proposta
          required: true
          schema:
            type: integer
            format: int64
        - name: rid
          in: path
          description: Identificador da revisão a atualizar
          required: true
          schema:
            type: integer
            format: int64
        - name: review
          in: query
          description: Objeto review que é atualizado
          required: true
          schema:
            $ref: '#/components/schemas/EventProposalReview'
      responses:
        '200':
          description: Revisão atualizada
        '400':
          description: Id especificado da proposta ou revisão inválidos
        '404':
          description: Proposta ou revisão não encontradas
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Event Proposal
      description: 'Remover do sistema a revisão identificada pelo {rid} da proposta {pid}'
      operationId: deleteProposalReview
      parameters:
        - name: pid
          in: path
          description: Identificador da proposta
          required: true
          schema:
            type: integer
            format: int64
        - name: rid
          in: path
          description: Identificador da revisão a remover
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Revisão removida
        '404':
          description: Porposta ou revisão não encontradas
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/proposals/{id}/comments':
    get:
      tags:
        - Event Proposal
      description: 'Obter todos os comentários da proposta de evento identificada pelo {id}'
      operationId: getEventProposalComments
      parameters:
        - name: id
          in: path
          description: Identificador da proposta de evento
          required: true
          schema:
            type: integer
            format: int64
        - name: criteria
          in: query
          description: Filtrar os comentários que satisfazem o critério
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Lista com os comentários da proposta de evento em questão
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventProposalComment'
        '404':
          description: Proposta de evento não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Event Proposal
      description: 'Adicionar um comentário à proposta de evento identificada pelo {id}'
      operationId: addEventProposalComment
      parameters:
        - name: id
          in: path
          description: Identificador da proposta de evento
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Todos os necessários dados do comentário
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventProposalComment'
      responses:
        '200':
          description: Comentário adicionado com sucesso
        '404':
          description: Proposta de evento não encontrada
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/proposals/{pid}/comments/{cid}':
    get:
      tags:
        - Event Proposal
      description: 'Obter o comentário identificado pelo {cid} na proposta {pid}'
      operationId: getProposalComment
      parameters:
        - name: pid
          in: path
          description: Identificador da proposta
          required: true
          schema:
            type: integer
            format: int64
        - name: cid
          in: path
          description: Identificador do comentário a obter
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Informação do comentário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventProposalComment'
        '404':
          description: Proposta ou comentário não encontrados
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Event Proposal
      description: 'Atualizar o comentário identificado pelo {cid} na proposta {pid}'
      operationId: updateProposalComment
      parameters:
        - name: pid
          in: path
          description: Identificador da proposta
          required: true
          schema:
            type: integer
            format: int64
        - name: cid
          in: path
          description: Identificador do comentário a atualizar
          required: true
          schema:
            type: integer
            format: int64
        - name: comment
          in: query
          description: Objeto comentário que é atualizado
          required: true
          schema:
            $ref: '#/components/schemas/EventProposalComment'
      responses:
        '200':
          description: Comentário atualizado
        '400':
          description: Id especificado da proposta ou comentário inválido
        '404':
          description: Proposta ou comentário não encontrados
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Event Proposal
      description: 'Remover do sistema o comentário identificado pelo {cid} na proposta {id}'
      operationId: deleteProposalComment
      parameters:
        - name: pid
          in: path
          description: Identificador da proposta
          required: true
          schema:
            type: integer
            format: int64
        - name: cid
          in: path
          description: Identificador do comentário a remover
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Comentário removido
        '404':
          description: Proposta ou comentário não encontrados
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      required:
        - firstName
        - lastName
        - userName
        - emailAddress
        - passwordHash
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        firstName:
          type: string
          example: john
        lastName:
          type: string
          example: smith
        userName:
          type: string
          example: user1
        emailAddress:
          type: string
          example: user1@ecma.com
        passwordHash:
          type: string
          example: $2y$12$95z6mYtPI371qXLCTWzP0elvLNv7Znh8dlwoseWncIm2zQb4u94xS
        approverInProposals:
          type: array
          items:
            type: number
...
    Employee:
      required:
        - userId
        - companyId
        - city
        - zipCode
        - streetAddress
        - cellPhone
        - homePhone
        - gender
        - salary
        - birthday
        - job
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        userId:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
        city:
          type: string
          example: Lisbon
        zipCode:
          type: string
          example: 1000-260
        streetAddress:
          type: string
          example: Short Street
        cellPhone:
          type: integer
          format: int64
          example: 919191919
        homePhone:
          type: integer
          format: int64
          example: 211111111
        gender:
          type: string
          enum:
            - M
            - F
        salary:
          type: integer
          format: int64
          example: 5000
          minimum: 0
        birthday:
          type: string
          example: '2018-01-01'
          format: date
        job:
          type: string
          example: CEO
        isAdmin:
          type: boolean
          default: false
    EventProposal:
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        isApproved:
          type: boolean
          default: false
        date:
          type: string
          format: date
          readOnly: true
        sections:
          type: array
          items:
            $ref: '#/components/schemas/EventProposalSection'
          readOnly: true
        reviewers:
          type: array
          items:
            $ref: '#/components/schemas/User'
          readOnly: true
        staff:
          type: array
          items:
            $ref: '#/components/schemas/User'
          readOnly: true
        members:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
          readOnly: true
        usersForBidding:
          type: array
          items:
            type: number
    EventProposalSection:
      required:
        - proposalId
        - title
        - description
        - goals
        - material
        - budget
        - workPlan
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        proposalId:
          type: integer
          format: int64
          readOnly: true
        title:
          type: string
          example: Section nr1
        description:
          type: string
          example: Section for a new perfume promotion event
        goals:
          type: string
          example: Boost sales by getting costumers to buy the product
        material:
          type: string
          example: Perfume samples
        budget:
          type: number
          format: float
          example: 5000
        workPlan:
          type: string
          example: >-
            Go to a place and promote the new perfume by allowing interested
            people to try it
    EventProposalReview:
      required:
        - proposalId
        - reviewerId
        - title
        - text
        - classification
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        proposalId:
          type: integer
          format: int64
          readOnly: true
        reviewerId:
          type: integer
          format: int64
          readOnly: true
        title:
          type: string
          example: Worst proposition ever
        text:
          type: string
          example: >-
            After analysing this proposition I came to the conclusion that this
            is not a very well-thought-out proposition because it lacks this and
            that
        summary:
          type: string
          example: A very bad proposition
        classification:
          type: integer
          format: int32
          minimum: 0
          maximum: 10
          example: 0
        date:
          type: string
          format: date
          example: '2018-10-10'
    EventProposalComment:
      required:
        - proposalId
        - commenterId
        - title
        - text
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        proposalId:
          type: integer
          format: int64
          readOnly: true
        commenterId:
          type: integer
          format: int64
          readOnly: true
        title:
          type: string
          example: Workplan question
        text:
          type: string
          example: >-
            Hello, can someone involved give a bit more detail about the
            workplan?
        date:
          type: string
          format: date
          example: '2018-10-15'
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string